Fix secureboot test regression due to output decoding
Many tests from tests/uefistored/test_secure_boot.py are failing with an
utf-8 decoding error happening in the VM.get_all_efi_bins method.

This is a regression introduced by commit
9524dcbc9a0d282f3f11f7b84d682b6815e1a09c which added a systematic
decoding of SSH command output. This doesn't play well will
VM.get_all_efi_bins which outputs the first two bytes of various binary
files, so this can't always be decoded as text.

Add a `decode` boolean parameter to the ssh functions, defaulting to True
but which when set to False will cause the output not to be decoded
(except for logging and error reporting, by setting errors='replace' in
order to be fault-tolerant).

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>