Format logs of failed SSH commands similarly to succeeded ones
The debug output of failed SSH commands currently begins with the error,
then the command and finishes with the output on a single line. This
renders not very well when several similar commands succeed then one
fails. It looks off.

Make it now the same as the normal debug output for SSH commands:
command, optional error code if needed, newline, output.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>