Add secureboot-certs --version check
This commit adds a check to the secureboot-certs version to ensure that
it is the correct version.

If secureboot-certs is an older version, the tests will fail in
mysterious ways (the certs will seemingly disappear without explanation,
due to the temp certs being named /tmp/tmp.XYZ instead of PK.auth,
KEK.auth, etc...).

Although this probably is unlikely to happen in the future, this short
fix seemed worth the potential hours lost in case tests get ran on a
host with an old uefistored package installed.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>