tests/uefistored: add db append test
This commit adds a test to test that DB appends succeed.

This should serve as a regression test for the bug that resulted in the
KB update error in Windows. Future work should perform that update in a
Windows VM, but this test ought to be a strong proxy for that
yet-to-be-supported feature of our test harness.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>