common: get_efi_var only returns that variable data
The Linux ABI for EFI variables returns both the attributes and the
variable data when reading from efivarfs.  The method get_efi_var()
subsequently was designed to return the attrs and the data. This ended
up being not used at all, so this commit changes it to only return the
used data.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>