lib/efi: refactor GUID strings and bytes into GUID class
Previously, EFI secure boot tests required passing around GUIDs that are
associated with EFI variables. The different tools for signing
variables, installing variables, and creating variables, all required
that the GUID be in different forms. Some require the GUID to be a hex
string, others require the GUID to be little-endian bytes.

The previous representation of GUIDs required that the caller knew what
form of GUID that callee required. For example, using set_efi_var()
required that you import EFI_GLOBAL_VARIABLE_GUID_STR (the string
representation) and pass it in, while Certificate.sign_data() required
the caller to import and pass in EFI_GLOBAL_VARIABLE_GUID (the bytes
representation). This was very inconvenient.

This commit combines the bytes and string representation into a GUID
class with a `as_str()` and a `as_bytes()` method. It also refactors
the functions that use the GUID to call the method associated with the
representation that they require. This allows GUID objects to be passed
around and used without knowing the internals of the callee.

For example, the previous mentioned functions now have a uniform
invocation:

>>> Certificate.sign_data(..., guid)
>>> vm.set_efi_var(..., guid)

Each function simply converts to string with as_str() or bytes with
as_bytes() as needed.

Note that Python's UUID already supports the two representations
required by GUID, which is why GUID simply inherits from it (which also
buys a proper __str__ and __repr__, and other helper methods). Ideally,
the methods could have been created as the property `GUID.bytes` and
function `str(GUID)` similar to the UUID class, but unfortunately
UUID.bytes_le calls UUID.bytes, so overriding GUID.bytes to call
self.bytes_le results in infinite recursion. As a compromise and for the
sake of uniformity, the as_str() and as_bytes() method names were chosen
instead.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>