Use scp to fix TestProxies tests
We can't use a json string argument on a host using host.xe
because we must already escape the simple/double quotes
of the ssh and xe commands. JSON introduces a new level of quotes...
Otherwise we have a bad shell interpretation of the quotes on the server.

Solution: we use scp to copy a script that contains the xe command on the host.

Signed-off-by: Ronan Abhamon <ronan.abhamon@vates.fr>