Initialize Pool object before Host
This solves an import cycle.

Now we don't create a Host object for master host then a Pool object that
itself creates Host objects for the other hosts in the pool. Instead, we
create the Pool object first (with the ip address or hostname of the
master host in the constructor's parameters) which then creates all the
Host objects. This also allows to get rid of the Host.initialize()
method after object creation.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>