Try to avoid failed sr-destroy due to incomplete GC
We sr-destroy runs, it triggers a single forced GC run, which may not
remove all hidden VDIs at once. Then sr-destroy fails due to SR not
being empty.

Try to detect this situation and retry several times in the hope that
the VDI chain that needs to be garbage collected will not be too long
when we need to destroy SRs.

Another method would be doing as suggested upstream: sr scan then wait
for GC to fully complete. But this can take a long time because the GC
process may enter its "quiet period" during which it does nothing.

See https://github.com/xapi-project/sm/issues/583

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>