Rename test SRs to make it obvious they're related to testing
I've seen cases when a test adds a SR with the same name as an already
existing SR that is not related to the tests. For some reason I then had
to manually remove the added SR after a test failure, and it was
confusing to have two SRs with the same name.

Add a -test suffix to the SR names created in our storage tests.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>