Rework moosefs tests where we remove moosefs-client
A single test would both test whether sr.scan fails if moosefs-client is
removed and whether attaching PBDs would fail after a reboot if
moosefs-client is missing.

The biggest issue with this test is that the reboot stalls because we
removed moosefs-client while the FS was still mounted.

I split the test and removed the reboot. Now there's one test that
checks sr.scan and another that tests re-attaching the PBD when moosefs-client
is missing, without waiting for a reboot to do so.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>