Fix option accumulation in _ssh()
Commit 0bec9b87a4d6df4c7f6dc9d13a7dd82e67dbf7e9 introduced a regression:
as the default value for parameter `options` is mutable, only one
reference is maintained by python, and so every call to _ssh() adds the
same parameters to the list, again and again, until the following
exception is raised: "OSError: [Errno 7] Argument list too long"

Use a copy of the parameter in the function to fix this.

For variable names consistency, the `options` variable in scp() is also
renamed to `opts`.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>