Remove broken set_proxies test and improve get_proxies test
The set_proxies test doesn't really test much currently as it only
attempts to remove repo proxies when hosts already don't have any proxies
set. The way it is implemented makes the test fail as soon as the list
of repositories in xcp-ng.repo is different from what is installed by
default, and attempting to improve the test reveals that set_proxies
doesn't have any practical use at the moment (setting a proxy will only
work if the repo URL is formatted in a special way that is specific to
apt-cacher-ng.

Let's remove the test altogether and add one back when/if the feature
becomes used by XO.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>