Mark tests as expected to fail in 8.3 or not relevant in 8.3
Tests marked as xfail are tests that are relevant in the context of
XCP-ng 8.3, currently fail, and will have to be fixed before the final
release. Some will be fixed when the behaviour of XAPI or varstored is
changed, and one will need to be modified to match a future new
behaviour (FIXME comment added).

One test was not relevant anymore, so I marked it as only relevant for
hosts less than 8.3.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>