SB tests: Remove all UEFI variables from imported VMs
In XCP-ng 8.2, with uefistored, removing just the SB certificates with
`varstore-sb-state $UUID setup` was enough to make it write certs to the
VM's NVRAM next time it boots.

In XCP-ng 8.3, with varstored, certificates are written in the NVRAM
store of a VM only if it has no UEFI variables at all.

In both cases, we now remove all UEFI variables from the VM after
importing it.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>