Split sb-main job to extract tests that require more than one host
In XCP-ng 8.2, the sb-main job could run on a single host, but it's
not the case in XCP-ng 8.3, as there are new tests covering certificate
propagation to all host disks at XAPI start, and a test covering
certificate propagation to a new host joining the pool. So the job needs
at least three hosts, two in a pool and one in a separate single-host pool.

In order to keep sb-main's job ability to run on a single host, the
tests in tests/uefistored/test_cert_inheritance.py are moved to a new
job called sb-certificates. It will still accept to run with a single
host on XCP-ng 8.2, but not on XCP-ng 8.3.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>