Rename xtf job and directory to xen
XTF is not the only way to test Xen, and it makes sense to have all
such tests in the same job and test directory.

So we rename xtf to xen, except the test_xtf.py file itself. This
prepares the way for adding new tests of the Xen hypervisor itself.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>