Don't skip missing name with yum install
When installing all packages together if one or more package don't exist
there is no error. Some tests are checking that packages can be installed
so we need to generate an error. To achieve this we add an option to yum
install to stop and generate an error if a package is missing.
Error is caught only for the first missing package.

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>