Add vtpm methods to VM class
This adds get_vtpm(), create_vtpm(), and destroy_vtpm() methods to the
VM class.

vm.get_vtpm() returns the vtpm uuid associated to the vm if it exists,
None otherwise.

vm.create_vtpm() creates a vtpm and associates it to the vm. It returns
the newly created vtpm uuid, or fails if a vtpm is already associated to
the vm.

vm.destroy_vm() destroys the existing vtpm associated to the vm if one
exists. It doesn't fail if no vtpm is associated to the vm.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>