Add force_if_fails parameter to vm.shutdown()
If vm.shutdown() is called with force_if_fails set to True and the first
call to 'xe vm-shutdown' fails, then a forced shutdown will be tried.
The default value for the parameter force_is_false is False. Also, force
and force_if_fails parameters are mutually exclusive and cannot be both
True.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>