Wait for 10 minutes before the 5th sr-destroy attempt
We already have code which retries a failed sr-destroy, when the cause is
remaining hidden VDIs which haven't been garbage-collected yet, and it's
enough in most cases, but not always.

Sometimes, the force-gc that is triggered by SM before the delete operation
is not enough. VDIs that look like they should be removed remain, and
are removed only a bit later, when the normal garbage collector runs.

This may be related to the fact that the forced GC operation doesn't
coalesce VDIs, or to a special case where a recently coalesced VDI is
temporarily protected against deletion.

We add a 10 minutes wait before the 5th sr-destroy attempt, to give the
normal garbage collector enough time to run (it should run 5 minutes
after the last operation that creates a need for the GC to run).

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>