Add sftp commands support
This patch adds the sftp() function used to send sftp commands to a
host. It takes an ip or a hostname as parameter and a list of sftp
commands.

sftp is executed in batchfile mode reading commands from the standard
input and connects to the host as user root.

Like the scp function, the sftp function supports the check and
suppress_fingerprint_warnings parameters to respectively check the
command return code and control the ssl fingerprint warnings.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>