Robustify Host.is_package_installed()
The current implementation is dependent on the exact output of the yum
list installed command, which can depend on the system language of the
user who runs the tests, and might change in the future.

Instead, use the simpler `rpm -q package` command, which returns nonzero
if the package doesn't exist on the system.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>