Avoid having two ISO SRs at the same time during tests
In ISO SR tests, there are package-scoped fixtures for SRs, which is
what we usually do in other storage tests. But here, in the same
package, we test ISO SR, NFS ISO SR, and CIFS ISO SR.

As a result we start creating a new SR when the previous one hasn't been
teared down yet.

By giving these fixtures the `module` scope, we ensure teardown happens
at the end of each test file. The counterpart is that we may create and
destroy the same SR twice, successively, because there are two test
files per SR type. However, since the second file is each time dedicated
to tests with reboot, and that we usually run these in a separate job,
this doesn't add any overhead to the current test jobs.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>