Don't skip empty lines...
For some tricky cases we cannot skip empty line. For example when
getting a key from varstored with ssh we can have within the key a
sequence of two 0x0a bytes. The issue is that readline will split it
with an empty line. If we skip it we basically remove one 0x0a from the
key and then of course the checksum will failed.

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>