Add data.py DEFAULT_SR setting
In some cases, we may prefer to favour a local SR to store test VM disks,
to avoid latency or unstabilities related to network or shared file servers.
However it's not good practice to make a local SR the default SR for a pool of several hosts.
Hence this configuration value that you can set to `local` so that our tests use this SR by default.
Possible values:
- 'default': keep using the pool's default SR
- 'local': use the first local SR found instead

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>