Allow to cache VMs in order to reuse them across test modules
A new parameter was added to data.py-dist and thus expected in data.py:

CACHE_IMPORTED_VM. Defaults to False.

Setting it to True will cache VMs imported via the `imported_vm` fixture.
When the VM is to be imported, we'll first look for an existing cache VM
on the same SR. This also causes the VM to be cloned at the beginning of
each test module, so that the original cached VM remains untouched.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>