Fix parameter order for BaseVM.param_set
key and value were inverted in this method, which is not consistent
with the definition of _param_set, nor with the othr param_set methods
on other classes.

This makes calls a bit awkward when a key must be specified in addition
to the parameter name, but it's consistent, mais we can make it
reasonably readable by specifying key as a named parameter in method
calls.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>