set_efi_var: Fix copy method of keys to the VM efivarfs
Keys used to be copied to the VM efivarfs using scp. This works when scp
uses its own legacy protocol but this doesn't work anymore when scp uses
the sftp protocol. To avoid future issues, this change modifies the way
keys are put to the VM by first copying them to the /tmp folder and then
cat them to the efivarfs (on Alpine images, cp from busybox is not able
to update keys in efivarfs so use 'cat tmp_key > key' instead).

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>