jobs.py: Add --host-version support for collect command
This patch adds the option --host-version (-v for short) to the 'jobs.py
collect' command allowing to filter VMs from vm_data.py based on a given
host version. This uses the same mechanism as for the run command but
the host version is specified instead of being obtained from the host
under test.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>