test_*_sr: fix test_snapshot (x)fail not to contaminate subsequent tests
All SR's test_snapshot take a stopped VM from vm_on_*_sr fixture.  Then
they start it, and hopefully stop it if everything went well.

Now if something went bad the shutdown would never be requested.  If
other tests are then run after a failure (which currently requires using
`--maxfail` to overrule the config file), or if such a test is marked
xfail (like the first public prototype for zfs-ng), subsequent tests chained
by pytest would be launched with a VM wrongly started, and fail for this
reason, which is bad.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>