Add tests for zfs-vol
Those are derived from the zfs tests.

Some tests were not ported, the expectation being we'll ship a
dedicated RPM for the SR driver, which will pull the zfs package,
which removes a whole bunch of trick situations.  We may want to write
tests for the case where the SR driver is not installed, but they will
be simpler.

Not all quicktests pass yet.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>