Host.reboot: use intermediate steps to better understand reboot delays
Code as showing "Wait for host up" while indeed waiting for XAPI to be up
and running.  This splits it between "Wait for host up" (== answers to
ping), "Wait for ssh up on host" (== port 22 open), and "Wait for XAPI to
be ready" (xapi says so over known-working ssh).

This notably allows to see some 1-minute delays between working IP and
open port.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>