uefi_sb: Test pool and VM cert states when no custom cert installed
We previously didn't check that a pool without custom certs only has
PK.auth in `/var/lib/varstored`. Check this.

We also didn't test what happens when you boot a fresh UEFI VM on a pool
which only has PK.auth, and this bit us. We thought it would behave some way,
and it actually behaved the other way. We changed it so that it behaves
as we need, so add a test which ensures this works and doesn't regress
in the future.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>