plugin_updater: fix pkgfixture usage
pkgfixture stuff must *always* be used from conftest.py, or else
strange problems appear because of the result shown in
https://github.com/pytest-dev/pytest/issues/8189#issuecomment-2158537878

For this case it results in a test failure when test ordering changes,
because of a safeguard added to the fixture.  And "stop using a test
parameter for sr_disk" commit does cause such a change in test
ordering.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>