install 4/n: make sure host appears in PXE ARP tables
Detection of host IP till now relies on the fact we download the
answerfile from PXE server.  Once we take this file from the ISO this
network traffic won't happen so we need some other mechanism to fill the
server's ARP tables.

test-pingpxe.service is installed in install.img by iso-remaster.

Since it is difficult to wait until the IP has been assigned before
launching the service, make it ping continuously until we can reach
the PXE server.

Similarly, once installed the host needs the same mechanism so the test
can find it.  We prepare an installer hook to install the service on the
host, which the generated answerfile will use in later commit.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>