Extend export/cache to work with classes
Context in https://github.com/pytest-dev/pytest/discussions/12541

Applies to a class with an autouse fixture dealing with host lifecycle,
and containing effect-less tests, where the cache applies to the whole
class and not to a single test.

At export time this adds the missing class-parameter set to the class
nodeid (like what pytest already does to functions).

A `test_anchor` is added to `continuation_of`, to generate a suitable
dependency to compensate for the lack of separate per-paramset class
nodes.