Add a first-boot service to make UUIDs unique
The installer generates UUID for host and its dom0 in xensource-inventory,
so the result of a costly install test could not be reused for multiple
hosts in a same pool.

This service will be run once and before xapi ever starts, to override
those UUIDs with brand new random ones during firstboot.