Add test for file system diff script
This test calls the script xcpng-fs-diff.py and fails if any difference
is found. It is very basic for now but can be improved if we want to
ignore some of the returned results. The script returns the results in a
json array to ease parsing.

The test needs 2 hosts and checks they are of the same XCP-ng version.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>