Image caching: allow to declare image equivalence
The key used to locate a VM image in the VM cache depends on the test repo
commit hash, to protect against using results from incompatible test by
mistake.  But the commit hash can change for many reasons that do not
influence the parent tests, so this provides a way to use known-equivalent
test outputs.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>