install 2/n: insert ISO in host VM
Getting a vm_booted_with_installer starts with inserting a CD, will be
progressively enhanced.

ISOs are specified through an iso_version key, matched to a filename,
or to a download URL used through a local cache, in data.py.

The ISO is uploaded to the SR through the pool master's mountpoint,
which will lead to extra network traffic but avoids having to
(mis)record the server's location.

The iso_version value is temporarily specified using a pytest marker.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>