install 7/n: answerfile generation
This process has several steps:
- building of a data structure holding all of the answerfile data,
  from a customizable base in data.py and from tests-specific items
- serialization as XML to be read by host-installer
- necessary changes to the ISO for host-installer to use it

We now have to explicitly enable the network during access (was implied by
the use of a remote answerfile).  Similarly we now rely on the test-pingpxe
service, as nothing else would otherwise populate the server's ARP table.

This is needed so:
- different tests can use different parameters without the need for
  provisionning every answerfile to be used
- tests can dynamically add contents for their own needs, before the
  XML gets actualy written

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>

use legacycrypt in place of crypt

crypt has been deprecated in python 3.11 and removed in 3.13

replacing crypt with another lib is not that easy. See

  https://eighty-twenty.org/2024/01/13/python-crypt-shacrypt

as an example. crypt_r is an alternative to legacycrypt that exports
the crypt module, but doesn't come with a binary installable package.

  https://pypi.org/project/crypt-r/

Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>