BaseVM: move implementation to VM
mypy will complain not finding _destroy in BaseVM once we add typing hints.

`Snapshot` and `VM` have had different implementations because of vm-uninstall
vs snapshot-uninstall, and diverged further because of xapi-project/xen-api#4145.
The code in `BaseVM.destroy` is specific to this issue's workaround for VM
and has no reason to be there.

This corrects the situation, and moves the comment about the xapi issue to
a better place.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>