setup mise
install `mise` once with

  curl https://mise.run | sh

then just run

  mise run test tests/misc/test_basic_without_ssh.py --hosts=gln-xcpng

to install all the required dependencies, including `xo-cli`, `python`,
and required python modules, and run the selected tests.

If you prefer not to use `mise run`, you can tell mise to install the
dependencies and then set up the shell environment to have them
accessible in the shell:

  mise install
  mise activate | source

`mise` deactivates itself as soon as you leave the project directory,
and reactivates when you enter a directory with a `mise` configuration,
so you may even put that last command in your global shell configuration.

Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>