setup uv
install `uv` once with one of these commands:

  pip install uv
  pipx install uv
  curl -LsSf https://astral.sh/uv/install.sh | sh

then just run

  uv run pytest tests/misc/test_basic_without_ssh.py --hosts=gln-xcpng

to install all the required python dependencies and run the selected
tests.

If you prefer not to use `uv run`, you can tell uv to install the
dependencies and then set up the shell environment to have them
accessible in the shell:

  uv sync
  source .venv/bin/activate

Run `deactivate` to deactivate the virtual environment.

Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>