feat(coalesce): Add tests for coalesce
Add a vdi-type parameter to parametrize the vdi_type fixture,
it default to vhd at the moment, will be used to run test on other vdi
type, e.g. qcow2.

The tests create a VDI, connect it to Dom0 then use the tapdev to write
random data somewhere in it.
It then compare the data of the VDI to the original data we have to see
if it has changed.
The test create snapshot/clone before deleting them and waiting for the
coalesce to have happened by observing sm-config:vhd-parent before
checking the integrity of the data.

Signed-off-by: Damien Thenot <damien.thenot@vates.tech>