WIP install: add ipv4/ipv6 variants
FIXME: ipv6 variant does not work as-is, as currently on firstboot the
host gets assigned an IPv4 at first, which causes the IP-detection
mechanism to use that IPv4 address, which quickly gets unusable to
communicate with the host.  Possibly (likely? or not?) comes from the
installer using DHCPv4 at first?