Stop pytest from always stopping at first error.
It is very annoying to get a single failed test stop the whole run by
default: it requires rerunning the whole run (or spending time forging
a filter to exclude exactly those we don't want to rerun), whereas
when a test has failed and the user wants to stop the run because of
that, Ctrl-C is our friend.

It is also very easy to just add `-x` to a pytest command to stop on
first error when really needed, much less obvious to dig for
`--maxfail=0` when we want to have a view of all errors.  Even in CI
it is not obvious that we always want this anyway, likely only for the
very long jobs.

And it is the standard behaviour, and will be the default expectation
of any newcomer with previous knowledge of pytest.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>