use flake8 in place of pycodestyle
`pycodestyle` is not supported anymore in vscode's python linter plugin,
making it difficult to run the style lint in vscode. `flake8` includes
`pycodestyle` and is better supported in vscode.

In addition, the vscode plugin `cornflakes-linter`, which uses `flake8`,
allows us to lint as we type, providing a much better experience.

The configuration is strictly identical to our pycodestyle configuration.

Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>