WIP host.reboot: fix detection of host up/down
When rebooting a master host, and no VM is running, it can get down fast
enough that the ssh connection is closed on us.  In this case (which we
already handled) trying to open a new ssh connection to query xapi is never
going to succeed. Even when a xapi master answers a slave is not enabled
it is a bit of a stretch to conclude the host is down already.

Especially, the test for "host up" uses ping, and the host might still
respond to a few pings before it actually shuts down.  Instead, for
robustness of the "host up" test, it is much more logical to use ping as
well to decide the host is down.