Host: use a single lsblk call to cache all static data about disks
Those info typically don't change over the course of a test session, but
Host.rescan_block_devices_info() is provided for when we'll have more
sophisticated test cases.

We use a custom TypedDict to be able to use this type more largely in
upcoming commits.

Reimplements disks() and available_disks() on top of this.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>