Robustify LINSTOR `_get_diskful_hosts` test regarding coalesce
It's possible that the test fails due to an incorrect verification
on the resource linked to a VDI. Indeed, if a coalesce occurs just
after fetching the diskless resource to be monitored, it may be
deleted which causes the test to fail. The fix is therefore to
refetch the resource associated with a VDI rather than caching it.

Signed-off-by: Ronan Abhamon <ronan.abhamon@vates.tech>